[frontend] refact
