+++ openbas-front/src/admin/components/components/channels/Channel.js
@@ -11,9 +11,9 @@ import ChannelParametersForm from './ChannelParametersForm';
 import ChannelOverviewNewspaper from './ChannelOverviewNewspaper';
 import ChannelOverviewMicroblogging from './ChannelOverviewMicroblogging';
 import ChannelOverviewTvChannel from './ChannelOverviewTvChannel';
-import ChannelAddLogo from './ChannelAddLogo';
 import useDataLoader from '../../../../utils/hooks/useDataLoader';
 import { fetchDocuments } from '../../../../actions/Document';
+import ChannelAddLogo from './ChannelAddLogo';
 
 const useStyles = makeStyles(() => ({
   root: {
