+++ openbas-front/src/admin/components/payloads/PayloadForm.js
@@ -12,7 +12,7 @@ import TagField from '../../../components/TagField';
 import PlatformField from '../../../components/PlatformField';
 import OldSelectField from '../../../components/fields/OldSelectField';
 import AttackPatternField from '../../../components/AttackPatternField';
-import DocumentLoader from '../../../components/fields/DocumentLoader';
+import FileLoader from '../../../components/fields/FileLoader';
 
 const useStyles = makeStyles(() => ({
   tuple: {
@@ -43,7 +43,7 @@ const PayloadForm = (props) => {
         requiredFields.push(...['file_drop_file']);
         break;
       default:
-      // do nothing
+            // do nothing
     }
     requiredFields.forEach((field) => {
       if (field === 'payload_platforms' && (!values[field] || values[field].length === 0)) {
@@ -67,10 +67,11 @@ const PayloadForm = (props) => {
         },
       }}
     >
-      {({ handleSubmit, form, values, submitting }) => (
+      {({ handleSubmit, form, values, submitting, dirty }) => (
         <form id="payloadForm" onSubmit={(event) => {
-          handleSubmit(event);
+          event.preventDefault();
           setSubmitButtonClicked(true);
+          handleSubmit(event);
         }}
         >
           <OldTextField
@@ -133,22 +134,22 @@ const PayloadForm = (props) => {
             </>
           )}
           {type === 'Executable' && (
-          <>
-            <DocumentLoader
-              name="executable_file"
-              label={t('Document')}
-              setFieldValue={form.mutators.setValue}
-              initialValue={values.executable_file}
-              InputLabelProps={{ required: true }}
-              onSubmit={submitButtonClicked}
-            />
-          </>
+            <>
+              <FileLoader
+                name="executable_file"
+                label={t('File')}
+                setFieldValue={form.mutators.setValue}
+                initialValue={values.executable_file}
+                InputLabelProps={{ required: true }}
+                onSubmit={submitButtonClicked}
+              />
+            </>
           )}
           {type === 'FileDrop' && (
             <>
-              <DocumentLoader
+              <FileLoader
                 name="file_drop_file"
-                label={t('Document')}
+                label={t('File')}
                 setFieldValue={form.mutators.setValue}
                 initialValue={values.file_drop_file}
                 InputLabelProps={{ required: true }}
@@ -190,12 +191,12 @@ const PayloadForm = (props) => {
                       style={{ marginTop: -2 }}
                       color="primary"
                     >
-                      <ControlPointOutlined />
+                      <ControlPointOutlined/>
                     </IconButton>
                     {meta.error && meta.touched && (
-                      <div className={classes.errorColor}>
-                        {meta.error}
-                      </div>
+                    <div className={classes.errorColor}>
+                      {meta.error}
+                    </div>
                     )}
                   </InputLabel>
                 </div>
@@ -238,7 +239,7 @@ const PayloadForm = (props) => {
                           size="small"
                           color="primary"
                         >
-                          <DeleteOutlined />
+                          <DeleteOutlined/>
                         </IconButton>
                       </ListItem>
                     );
@@ -267,12 +268,12 @@ const PayloadForm = (props) => {
                       style={{ marginTop: -2 }}
                       color="primary"
                     >
-                      <ControlPointOutlined />
+                      <ControlPointOutlined/>
                     </IconButton>
                     {meta.error && meta.touched && (
-                      <div className={classes.errorColor}>
-                        {meta.error}
-                      </div>
+                    <div className={classes.errorColor}>
+                      {meta.error}
+                    </div>
                     )}
                   </InputLabel>
                 </div>
@@ -324,7 +325,7 @@ const PayloadForm = (props) => {
                           size="small"
                           color="primary"
                         >
-                          <DeleteOutlined />
+                          <DeleteOutlined/>
                         </IconButton>
                       </ListItem>
                     );
@@ -388,7 +389,7 @@ const PayloadForm = (props) => {
               variant="contained"
               color="secondary"
               type="submit"
-              disabled={submitting}
+              disabled={submitting || !dirty}
             >
               {editing ? t('Update') : t('Create')}
             </Button>
